Equip Items Added

+++ Assets/Scripts/BreakableObject.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class BreakableObject : MonoBehaviour, IDamageable
 {
-    [SerializeField] int hp = 10;
+    [SerializeField] float hp = 10;
 
-    public void ApplyDamage(int damage)
+    public void ApplyDamage(float damage)
     {
         hp -= damage;
     }
 
-    public void CalculateDamage(ref int damage)
+    public void CalculateDamage(ref float damage)
     {
         
     }
